Add support for express middleware.
As it stood before this change, express middleware couldn't be added
because it would be loaded after the route handlers.

This makes it possible for you to add middleware:

```js
'use strict';
const BootBot = require('bootbot');

// Bot

const bot = new BootBot({
  accessToken: process.env.FACEBOOK_PAGE_TOKEN,
  verifyToken: process.env.FACEBOOK_VERIFY_TOKEN,
  appSecret:   process.env.FACEBOOK_APP_SECRET,
});

bot.app.use(function(req, _res, next) {
  console.log("In middleware")

  // Mutate first message.

  let msg = req.body.entry[0].messaging[0].message
  msg.text = msg.text.toUpperCase() // OMG!!!!

  next()
})

bot.start(420)

```

The tests are passing.

This could be an possible alternative to the proposal in #3. Or at least
a temporary alternative.

Cheers!!! <3 <3 <3