Don't remove buttons without title in _formatButtons
The _formatButton method, called by sendButtonTemplate and sendListTemplate,
discarded all buttons without a 'title' key.
But not all buttons usable in those methods are required to have a title:
notably, the login and logout buttons of the account linking flow must
only specify the 'type' and 'url' keys.

This resulted in an error when trying to send a button template with only
a login button, for example using:

  bot.say({
    text: 'In order to use the bot, you must authenticate to our website',
    buttons: [{type: 'account_link', url: '<myLoginUrl>'}]
  })

because the bot *silently* discarded the only button, causing
Facebook to return an error due to a request for a button template message
without buttons.